amd64: fix compilation issues

+++ ivp_compact_builder/ivp_compact_recursive.cxx
@@ -63,7 +63,7 @@ void IVP_Compact_Recursive::build_convex_hull(){
 
 void IVP_Compact_Recursive::set_rekursive_convex_hull(){
   IVP_Hash point_hash( 1024, sizeof(IVP_U_Float_Point),(void *)-1);
-  int n_points_in_hash = 0;
+  intp n_points_in_hash = 0;
   
   struct Triangle_Key {
     int point_index[3];
