+++ ivp_compact_builder/ivp_convex_decompositor.cxx
@@ -9,7 +9,7 @@
 
 
 
-void IVP_Concave_Polyhedron_Face::add_offset(int offset_in) {
+void IVP_Concave_Polyhedron_Face::add_offset(intp offset_in) {
 
     IVP_Concave_Polyhedron_Face_Pointoffset *p_offset = new IVP_Concave_Polyhedron_Face_Pointoffset();
     p_offset->offset = offset_in;
@@ -51,11 +51,11 @@ class IVP_Convex_Subpart_Work {
 public:
     IVP_U_Vector<int> offsets; // @@@SF: replace with IVP_VHash someday!!!
 
-    void add_offset(int offset_in);
+    void add_offset(intp offset_in);
 };
 
 
-void IVP_Convex_Subpart_Work::add_offset(int offset_in) {
+void IVP_Convex_Subpart_Work::add_offset(intp offset_in) {
 
     int i;
     for (i=0; i<this->offsets.len(); i++) {
